+++ src/cc/include/swcdb/fs/HadoopJVM/FileSystem.h
@@ -72,11 +72,11 @@ class FileSystemHadoopJVM: public FileSystem {
   SmartFdHadoopJVM::Ptr get_fd(SmartFd::Ptr &smartfd);
 
   void write(int &err, SmartFd::Ptr &smartfd,
-             int32_t replication, int64_t blksz, 
+             uint8_t replication, int64_t blksz, 
              StaticBuffer &buffer) override;
 
   void create(int &err, SmartFd::Ptr &smartfd, 
-              int32_t bufsz, int32_t replication, int64_t blksz) override;
+              int32_t bufsz, uint8_t replication, int64_t blksz) override;
 
   void open(int &err, SmartFd::Ptr &smartfd, int32_t bufsz = -1) override;
